# backend/docs/social-api-draft.openapi.yaml
...
openapi: 3.0.3
info:
  title: STROITECH Social API (draft)
  version: "0.1.0"
  description: >
    Планируемые маршруты. Префикс API в проде обычно /api (как у текущих /api/users, /api/feed).
servers:
  - url: /api
...
paths:
...
  /events/reminders/dispatch:
    post:
      tags: [events]
      summary: Отправка in-app напоминаний участникам со статусом «иду» (событие начинается в ближайшие 24 ч)
      description: Заголовок X-Cron-Secret должен совпадать с EVENT_REMINDER_CRON_SECRET; без переменной окружения — 503.
      parameters:
        - name: X-Cron-Secret
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK (data.dispatched — число отправленных)
        "403":
          description: Неверный секрет
        "503":
          description: Не настроено
  /events:
    get:
      tags: [events]
      summary: Список событий (предстоящие / прошедшие)
      parameters:
        - name: when
          in: query
          schema: { type: string, enum: [upcoming, past], default: upcoming }
        - name: communityId
          in: query
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: OK
    post:
      tags: [events]
      summary: Создать событие (при communityId — нужно быть участником сообщества)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [events]
      summary: Карточка + группы участников; myRsvp и canManage при JWT
      responses:
        "200":
          description: OK
    put:
      tags: [events]
      summary: Изменить (создатель, staff сообщества или глобальный moderator)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    delete:
      tags: [events]
      summary: Удалить
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /events/{id}/rsvp:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [events]
      summary: RSVP (иду / возможно / не иду); для события сообщества — только участники
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [going, maybe, not_going] }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...